% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_output.R
\name{calc_SSB}
\alias{calc_SSB}
\alias{calc_biomass}
\title{Calculate Spawning Stock Biomass (SSB) or total biomass}
\usage{
calc_SSB(wgt, mature, N)

calc_biomass(wgt, N)
}
\arguments{
\item{wgt}{A matrix with dimensions \code{nsc} and \code{nfish} representing the weight of each species in each length class.}

\item{mature}{A matrix with dimensions \code{ncs} and \code{nfish} with elements in the range 0-1 representing the proportion of mature individuals of each species in each length class.}

\item{N}{A matrix with dimensions \code{nsc} and \code{nfish} representing the number of individuals in each length class for the current time step.}
}
\value{
\code{calc_SSB} returns a numeric vector of length \code{nfish} representing the SSB of each species (g).

\code{calc_biomass} returns a numeric vector of length \code{length(species)} where the \code{j}th element is the biomass (g) of the \code{j}th species.
}
\description{
Calculates the spawning stock biomass (SSB) or total biomass of each species in the model.
}
\details{
The Spawning Stock Biomass is equal to \code{colSums(mature*N*wgt)} and the total biomass is equal to \code{colSums(mature*N*wgt)}.
}
\examples{
# Set up and run the model
NS_params <- LeMansParam(NS_par, tau=NS_tau, eta=rep(0.25, 21), L50=NS_par$Lmat, other=1e12)
effort <- matrix(0.5, 10, dim(NS_params@Qs)[3])
model_run <- run_LeMans(NS_params, years=10, effort=effort)

# Calculate SSB
calc_SSB(wgt=NS_params@wgt, mature=NS_params@mature, N=model_run@N[,,101])

# Calculate biomass in the last time step
calc_biomass(wgt=NS_params@wgt, N=model_run@N[,,101])
}
\seealso{
\code{\link{calc_recruits}}
}
