% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_LeMans.R
\docType{class}
\name{LeMans_outputs-class}
\alias{LeMans_outputs-class}
\title{An S4 class representing the outputs of the LeMans model}
\description{
An S4 class representing the outputs of the LeMans model
}
\section{Slots}{

\describe{
\item{\code{N}}{An array with dimensions \code{nsc}, \code{nfish} and \code{tot_time} representing the number of individuals in each length class at each time step of the model.}

\item{\code{Catch}}{An array with dimensions \code{nsc}, \code{nfish} and \code{tot_time} representing the biomass of each species that is caught in each length class at each time step of the model.}

\item{\code{M2}}{An array with dimensions \code{nsc}, \code{nfish} and \code{tot_time} representing the predation mortality of each species in each length class at each time step of the model.}

\item{\code{R}}{A matrix with dimensions \code{tot_time} and \code{nfish} representing the number of recruits of each species at each time step of the model.}
}}

