% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NS_par.R
\docType{data}
\name{NS_eta}
\alias{NS_eta}
\title{The steepness of the slope of the catchability curve}
\format{A numeric value representing the steepness of the slope of the catchability curve.}
\usage{
NS_eta
}
\description{
The steepness of the slope of the catchability curve associated with \code{NS_mixed_fish}.
}
\references{
Thorpe, R.B., Dolder, P.J. , Reeves, S., Robinson, P., Jennings, S. (2015). Assessing fishery and ecological consequences of alternative management options for multispecies fisheries \emph{ICES Journal of Marine Science}, 73(6):1503-1512.
}
\keyword{datasets}
