% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NS_par.R
\docType{data}
\name{NS_other}
\alias{NS_other}
\title{Other food for the North Sea}
\format{A numerical value representing other food. To be used with \code{NS_par}.}
\usage{
NS_other
}
\description{
Other food for the North Sea dataset, \code{NS_par}.
}
\keyword{datasets}
