% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NS_par.R
\docType{data}
\name{NS_tau}
\alias{NS_tau}
\title{North Sea interaction matrix}
\format{A matrix with 21 rows and 21 columns:
\describe{
 Row indices represent predators and column indices represent prey. A value of 1 at location \code{i}, \code{j} indicates prey \code{j} is eaten by predator \code{i}.
}}
\usage{
NS_tau
}
\description{
A predator-prey interaction matrix for the 21 species in \code{NS_par}.
}
\references{
Thorpe, R.B., Le Quesne, W.J.F., Luxford, F., Collie, J.S., Jennings, S. (2015). Evaluation and management implications of uncertainty in a multispecies size-structured model of population and community responses to fishing. \emph{Methods in Ecology and Evolution}, 6:49-58.
}
\keyword{datasets}
